shuffling gridding routines.
